Create subcomponent
